cleaned up Loader and created a reusable bouncing SVG component
