+++ src/components/Loader/data.js
@@ -1,0 +1,20 @@
+export const LOADER_DATA = [
+    { 
+        id: 0,
+        color: 'pink-600',
+        wrapperColor: 'pink-500',
+        bounceAmount: ''
+    },
+    { 
+        id: 1,
+        color: 'yellow-400',
+        wrapperColor: 'yellow-200',
+        bounceAmount: '200'
+    },
+    { 
+        id: 2,
+        color: 'green-400',
+        wrapperColor: 'green-300',
+        bounceAmount: '400'
+    }
+];
