+++ src/components/hoc/auth/ProtectedRoute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 import { withAuthenticationRequired } from '@auth0/auth0-react';
-import Loader from '../../Loader';
+import Loader from '../../Loader/Loader';
 
 const ProtectedRoute = ({ component, ...args }) => (
   <Route
